requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
fonttools==4.44.0
frozenlist==1.4.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.1
multidict==6.0.4
networkx==3.1
numpy==1.26.0
openai==0.28.1
packaging==23.2
pandas==2.1.2
Pillow==10.1.0
pymongo==4.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2
